Redo database schema for simplicity and graphql (#399)
This is going to remove many features from PlateZero and really bring it
back to the core focus of storing recipes, updating them and writing
notes for yourself for next time.

This new database is going to use [Graphile][] as a GraphQL layer to
automatically create the schema. We will still use JWT tokens to manage
user authentication. We will use Postgres Roles and security access
layers to manage what the user can do.

Create two schemas, one public and one private. GraphQL can only access
the public schema. Further we use the authorization to limit what the
user can do.

Start layering in comments, which turn into GraphQL documentation.

This has the basis of registering a user and
authentication/authorization.

Update to Postgres 13.

[Graphile]: https://www.graphile.org/

* Add Apollo; add Register Mutation
* Remove routes file

This is no longer needed without a custom server. We will use Next.js's
built in routing.

* Fix prettier