Updated User Page + Recipe Creation
Barebones implementations for a new user page and creating a recipe, all
hard coded for now. Working on getting the basics down with GraphQL and
updating the database to ensure users can only update and insert when
they are logged in.

Removed and simplified several pages that add complexity. Changing over
several old pages to be functional and use hooks.